v3.1 - adding support
