+++ aiopenapi3/__init__.py
@@ -1,7 +1,7 @@
+from .version import __version__
 from .openapi import OpenAPI
 from .loader import FileSystemLoader
 from .errors import SpecError, ReferenceResolutionError
-from .version import __version__
 
 
 __all__ = ["__version__", "OpenAPI", "FileSystemLoader", "SpecError", "ReferenceResolutionError"]
