+++ aiopenapi3/v30/general.py
@@ -1,10 +1,8 @@
-import urllib.parse
 from typing import Optional
 
 from pydantic import Field, Extra
-from yarl import URL
 
-from .object_base import ObjectExtended, ObjectBase
+from ..base import ObjectExtended, ObjectBase
 
 
 class ExternalDocumentation(ObjectExtended):
@@ -17,26 +15,6 @@ class ExternalDocumentation(ObjectExtended):
 
     url: str = Field(...)
     description: Optional[str] = Field(default=None)
-
-
-class JSONPointer:
-    @staticmethod
-    def decode(part):
-        """
-
-        https://swagger.io/docs/specification/using-ref/
-        :param part:
-        """
-        part = urllib.parse.unquote(part)
-        part = part.replace("~1", "/")
-        return part.replace("~0", "~")
-
-
-class JSONReference:
-    @staticmethod
-    def split(url):
-        u = URL(url)
-        return str(u.with_fragment("")), u.raw_fragment
 
 
 class Reference(ObjectBase):
