+++ aiopenapi3/v30/servers.py
@@ -2,7 +2,7 @@ from typing import List, Optional, Dict
 
 from pydantic import Field
 
-from .object_base import ObjectExtended
+from ..base import ObjectExtended
 
 
 class ServerVariable(ObjectExtended):
