+++ aiopenapi3/v30/tag.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pydantic import Field
 
-from .object_base import ObjectExtended
+from ..base import ObjectExtended
 from .general import ExternalDocumentation
