+++ aiopenapi3/v31/__init__.py
@@ -1,0 +1,5 @@
+# from .schemas import Schema
+from .root import Root
+
+# from .paths import PathItem, Operation, SecurityRequirement
+# from .parameter import Parameter
