+++ aiopenapi3/v31/media.py
@@ -2,9 +2,10 @@ from typing import Union, Optional, Dict, Any
 
 from pydantic import Field
 
+from ..base import ObjectExtended
+
 from .example import Example
 from .general import Reference
-from .object_base import ObjectExtended
 from .schemas import Schema
 
 
@@ -12,7 +13,7 @@ class Encoding(ObjectExtended):
     """
     A single encoding definition applied to a single schema property.
 
-    .. _Encoding: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#encoding-object
+    .. _Encoding: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#encodingObject
     """
 
     contentType: Optional[str] = Field(default=None)
@@ -27,7 +28,7 @@ class MediaType(ObjectExtended):
     A `MediaType`_ object provides schema and examples for the media type identified
     by its key.  These are used in a RequestBody object.
 
-    .. _MediaType: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.0.3.md#media-type-object
+    .. _MediaType: https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#media-type-object
     """
 
     schema_: Optional[Union[Schema, Reference]] = Field(required=True, alias="schema")
