+++ aiopenapi3/v31/security.py
@@ -1,0 +1,1 @@
+from ..v30.security import SecurityScheme, SecurityRequirement
