+++ aiopenapi3/v31/tag.py
@@ -1,0 +1,1 @@
+from ..v30.tag import Tag
