+++ aiopenapi3/v31/xml.py
@@ -1,0 +1,1 @@
+from ..v30.xml import XML
