+++ aiopenapi3/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.1"
+__version__ = "0.1.2"
