+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = aiopenapi3
-version = attr: aiopenapi3.__version__
+version = attr: aiopenapi3.version.__version__
 url = https://github.com/commonism/aiopenapi3
 description = OpenAPI3 3.0.3 client / validator based on pydantic & httpx
 long_description = file: README.md
@@ -48,6 +48,3 @@ tests =
     fastapi-versioning
     hypercorn
     uvloop
-    tatsu
-expression =
-    tatsu
